requirements.txt
git+git://github.com/django-nonrel/django
git+git://github.com/django-nonrel/djangoappengine.git
git+git://github.com/adieu/django-autoload
git+git://github.com/django-nonrel/django-dbindexer
djangorestframework==3.1.3
django-filter
markdown
git+git://github.com/django-nonrel/djangotoolbox
